+++ braintree/customer.py
@@ -162,6 +162,7 @@ class Customer(Resource):
             "device_session_id", "fraud_merchant_id", # NEXT_MAJOR_VERSION remove device_session_id and fraud_merchant_id
             {"risk_data": ["customer_browser", "customer_device_id", "customer_ip", "customer_location_zip", "customer_tenure"]},
             {"credit_card": CreditCard.create_signature()},
+            {"apple_pay_card": ApplePayCard.signature()},
             {"custom_fields": ["__any_key__"]},
             {"three_d_secure_pass_thru": [
                 "cavv",
@@ -186,6 +187,7 @@ class Customer(Resource):
         return [
             "company", "email", "fax", "first_name", "id", "last_name", "phone", "website", "device_data", "device_session_id", "fraud_merchant_id", "payment_method_nonce", "default_payment_method_token",
             {"credit_card": CreditCard.signature("update_via_customer")},
+            {"apple_pay_card": ApplePayCard.signature()},
             {"three_d_secure_pass_thru": [
                 "cavv",
                 "ds_transaction_id",